RAG_Package/requirements.txt
pymilvus>=2.3.0
llama-index-core>=0.9.0
llama-index-embeddings-huggingface>=0.1.0
transformers>=4.30.0
torch>=2.0.0 
llama-index-readers-file>=0.4.7
torchvision>=0.22.0
jsonpath-ng>=1.7.0